refactor(mainnet): add pallet_utility to utility
